GraphQL server refractor
