+++ litmus-portal/graphql-server/pkg/analytics/tsdbs/prometheus/prometheus.go
@@ -10,11 +10,12 @@ import (
 	"time"
 
 	"github.com/jinzhu/copier"
-	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/graph/model"
-	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/graphql"
-	api "github.com/prometheus/client_golang/api"
+	"github.com/prometheus/client_golang/api"
 	apiv1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	md "github.com/prometheus/common/model"
+
+	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/graph/model"
+	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/types"
 )
 
 func CreateClient(url string) (apiv1.API, error) {
@@ -33,7 +34,7 @@ func CreateClient(url string) (apiv1.API, error) {
 	return apiv1.NewAPI(client), nil
 }
 
-func Query(prom graphql.PromQuery) (model.PromResponse, error) {
+func Query(prom types.PromQuery) (model.PromResponse, error) {
 	client, err := CreateClient(prom.URL)
 	if err != nil {
 		return model.PromResponse{}, err
@@ -67,20 +68,20 @@ func Query(prom graphql.PromQuery) (model.PromResponse, error) {
 	}
 
 	var (
-		newResponse Response
-		newTSVs     [][]*TimeStampValue
+		newResponse types.Response
+		newTSVs     [][]*types.TimeStampValue
 		newLegends  [][]*string
 	)
 
 	for _, v := range data {
 
 		var (
-			tempTSV     []*TimeStampValue
+			tempTSV     []*types.TimeStampValue
 			tempLegends []*string
 		)
 
 		for _, value := range v.Values {
-			temp := &TimeStampValue{
+			temp := &types.TimeStampValue{
 				Timestamp: func(str string) *string { return &str }(fmt.Sprint(value.Timestamp)),
 				Value:     func(str string) *string { return &str }(fmt.Sprint(value.Value)),
 			}
@@ -121,15 +122,4 @@ func Query(prom graphql.PromQuery) (model.PromResponse, error) {
 	copier.Copy(&resp, &newResponse)
 
 	return resp, nil
-}
-
-type Response struct {
-	Queryid string              `json:"queryid"`
-	Legends [][]*string         `json:"legends"`
-	Tsvs    [][]*TimeStampValue `json:"tsvs"`
-}
-
-type TimeStampValue struct {
-	Timestamp *string `json:"timestamp"`
-	Value     *string `json:"value"`
 }
