+++ litmus-portal/graphql-server/pkg/database/mongodb/operations/gitops.go
@@ -14,8 +14,7 @@ import (
 )
 
 var (
-	gitOpsCollection  *mongo.Collection
-	backgroundContext = context.Background()
+	gitOpsCollection *mongo.Collection
 )
 
 const timeout = 15 * time.Second
