+++ litmus-portal/graphql-server/pkg/types/graphql.go
@@ -1,4 +1,4 @@
-package graphql
+package types
 
 type PromQuery struct {
 	Queryid    string
